+++ wdl/GermlineCNVTasks.wdl
@@ -82,6 +82,7 @@ task FilterIntervals {
       File? gatk4_jar_override
       String gatk_docker
       RuntimeAttr? runtime_attr_override
+      Boolean shared_filesystem = false
     }
 
     RuntimeAttr default_attr = object {
@@ -105,7 +106,18 @@ task FilterIntervals {
         export GATK_LOCAL_JAR=~{default="/root/gatk.jar" gatk4_jar_override}
 
         read_count_files_list=~{write_lines(read_count_files)}
-        grep gz$ $read_count_files_list | xargs -l1 -P0 gunzip
+        # Adding this for FSx/local FS 
+        # If the file is compressed, then only uncompressed it.
+        if [ ~{shared_filesystem} ]; 
+        then
+            while read file; do
+                if [[ $file == *.gz ]]; then
+                    if [ ! -f "${file%.gz}" ]; then gunzip -k $file;else echo "done";fi
+                fi
+            done < $read_count_files_list
+        else
+            grep gz$ $read_count_files_list | xargs -l1 -P0 gunzip
+        fi
         sed 's/\.gz$//' $read_count_files_list | \
             awk '{print "--input "$0}' > read_count_files.args
 
@@ -343,6 +355,7 @@ task BundledPostprocessGermlineCNVCalls {
         File? gatk4_jar_override
         String gatk_docker
         RuntimeAttr? runtime_attr_override
+        Boolean shared_filesystem = false
     }
 
     Int disk_gb_baseline = 10
@@ -385,7 +398,11 @@ task BundledPostprocessGermlineCNVCalls {
 
         mkdir -p extracted-contig-ploidy-calls
         tar xzf ~{contig_ploidy_calls_tar} -C extracted-contig-ploidy-calls
-        rm ~{contig_ploidy_calls_tar}
+        # Adding this for FSx/local FS 
+        if [ ! ~{shared_filesystem} ]; 
+        then
+          rm ~{contig_ploidy_calls_tar}
+        fi
 
         allosomal_contigs_args="--allosomal-contig ~{sep=" --allosomal-contig " allosomal_contigs}"
 
@@ -433,6 +450,7 @@ task PostprocessGermlineCNVCalls {
         File? gatk4_jar_override
         String gatk_docker
         RuntimeAttr? runtime_attr_override
+        Boolean shared_filesystem = false
     }
 
     Int disk_gb_baseline = 10
@@ -492,7 +510,11 @@ task PostprocessGermlineCNVCalls {
 
         mkdir -p contig-ploidy-calls
         tar xzf ~{contig_ploidy_calls_tar} -C contig-ploidy-calls
-        rm ~{contig_ploidy_calls_tar}
+        # Adding this for FSx/local FS 
+        if [ ! ~{shared_filesystem} ]; 
+        then
+          rm ~{contig_ploidy_calls_tar}
+        fi
 
         allosomal_contigs_args="--allosomal-contig ~{sep=" --allosomal-contig " allosomal_contigs}"
